[enhancement/3] Move DayCard to WeekRow
